# Introduce cuFFT plan cache; switch to auto-managed memory.