This PR introduces a cache for CUFFT handles, much like how cupy recently did.

Before:

```
julia> @benchmark fft($A)
BenchmarkTools.Trial: 1166 samples with 1 evaluation.
 Range (min … max):  2.479 ms … 281.481 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.999 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.274 ms ±   8.145 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

                    ▃ ▁  ▁ ▂▄█▇▃▃▅▂▄▃▃ ▁ ▁▂▃▂▅ ▁  ▁
  ▃▁▁▁▂▁▂▃▁▃▄▃▆▇▇████▇██▆███████████████▇███████▇██▇▇▇▄▅▄▃▃▃▃ ▅
  2.48 ms         Histogram: frequency by time        5.26 ms <

 Memory estimate: 2.81 KiB, allocs estimate: 64.
```

After:

```
julia> @benchmark fft($A)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   84.129 μs … 805.442 μs  ┊ GC (min … max):  0.00% … 91.37%
 Time  (median):     275.457 μs               ┊ GC (median):    87.83%
 Time  (mean ± σ):   276.244 μs ±   9.010 μs  ┊ GC (mean ± σ):  87.78% ±  0.98%

                                                           █▂
  ▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂██▄▂ ▂
  84.1 μs          Histogram: frequency by time          286 μs <
```

Reference:

```
❯ python wip.py
fft_func            :    CPU:   94.263 us   +/-16.328 (min:   68.750 / max:  421.615) us     GPU-0:  110.159 us   +/-31.509 (min:   42.400 / max: 2295.744) us
```

Closes https://github.com/JuliaGPU/CUDA.jl/issues/1682.